Make test_kill_scheduler check scheduler relaunched
